requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0
openai>=1.0.0
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.7.0
datasets>=2.14.0
evaluate>=0.4.0
accelerate>=0.20.0

# Document parsing & OCR
pypdf>=3.12.0
python-docx>=1.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow>=10.0.0


